+++ src/openssl_wrap.cpp
@@ -333,7 +333,7 @@ void _X509_REQ_set_version(X509_REQ* req, unsigned long version)
     OpensslCallIsOne::callChecked(lib::OpenSSLLib::SSL_X509_REQ_set_version, req, version);
 }
 
-X509_NAME* _X509_REQ_get_subject_name(X509_REQ *req)
+X509_NAME* _X509_REQ_get_subject_name(const X509_REQ *req)
 {
     return OpensslCallPtr::callChecked(lib::OpenSSLLib::SSL_X509_REQ_get_subject_name, req);
 }